+++ Sources/RepoBarCore/Support/Constants.swift
@@ -22,4 +22,5 @@ public enum RepoDetailCacheConstants {
     public static let trafficTTL: TimeInterval = 60 * 60
     public static let heatmapTTL: TimeInterval = 60 * 60
     public static let releaseTTL: TimeInterval = 60 * 60
+    public static let discussionsCapabilityTTL: TimeInterval = 24 * 60 * 60
 }